+++ src/fabric_bps/collectors/__init__.py
@@ -6,6 +6,7 @@ from .base import (
 from .capacities import collect_capacities
 from .domains import collect_domains
 from .pipelines import collect_pipelines
+from .scanner import collect_scanner_items, flatten_scan_result
 from .tenant_settings import collect_tenant_settings
 from .workspaces import collect_workspaces
 
@@ -16,6 +17,8 @@ __all__ = [
     "collect_capacities",
     "collect_domains",
     "collect_pipelines",
+    "collect_scanner_items",
+    "flatten_scan_result",
     "collect_tenant_settings",
     "collect_workspaces",
 ]